% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArtistID.R
\name{getArtistID}
\alias{getArtistID}
\title{Retrieves an artist ID from a general name search query}
\usage{
getArtistID(artist = NA, authentication_token = getAuthenticationToken())
}
\arguments{
\item{artist}{The artist to search for as a string}

\item{authentication_token}{Predefined argument which runs getAuthenticationToken()}
}
\value{
Dataframe containing the artist name, ID, popularity, and related genres
}
\description{
Retrieves an artist ID from a general name search query
}
\examples{
getArtistID('alt j')
}
