% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EricsFunctions.R
\name{getArtistInfo}
\alias{getArtistInfo}
\title{Get Data on a Spotify Artist}
\usage{
getArtistInfo(
  artist,
  byName = FALSE,
  dataframe = TRUE,
  lim = 10,
  authentication_token = auth_token
)
}
\arguments{
\item{artist}{The id or name of a Spotify artist to search.}

\item{byName}{Boolean. TRUE searches artist by name, FALSE searches by id.}

\item{dataframe}{Boolean. TRUE returns data cleaned in a dataframe, FALSE returns raw json.}

\item{lim}{The number of results to return if searched by name.}

\item{authentication_token}{The users authentication token. Defaults to the value returned by get_authentication_token.}
}
\value{
A dataframe or json object containing an artists information.
}
\description{
Search using either an artists name or Spotify id code.
}
\examples{
getArtistInfo("Ghost", byName = TRUE, dataframe = TRUE)

}
