% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_InfoAPI.R
\name{getAudioFeatures}
\alias{getAudioFeatures}
\title{Get Audio Features of a Song}
\usage{
getAudioFeatures(
  songId,
  output = "dataframe",
  authentication_token = get_authentication_token(CLIENT_ID =
    "7870a259411b4c8b8d2ad173b5a7ed73", CLIENT_SECRET =
    "67ab42b91f224c3682ff8d5b2220f6aa")
)
}
\arguments{
\item{songId}{The Id of a song to search}

\item{output}{Output type. "json", "dataframe", or "graph"}

\item{authentication_token}{the users authentication token for the Spotify API. Defaults to the value saved when get_authentication_token is run.}
}
\value{
A json object, dataframe, or graph describing song features
}
\description{
Spotify automatically generates a number of audio features for each song.
}
\examples{
getAudioFeatures("68ngtC3pGiTjXcFwxYCJ7Z", output = "graph")
}
