% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_PodcastAPI.R
\name{getAuthenticationToken}
\alias{getAuthenticationToken}
\title{Authentication Token}
\usage{
getAuthenticationToken(client_id, client_secret_id)
}
\arguments{
\item{client_id}{string: (Visit https://developer.spotify.com/dashboard for more information)}

\item{client_secret_id}{string: (Visit https://developer.spotify.com/dashboard for more information)}
}
\value{
authentication token saved to a global variable authentication_token
}
\description{
Get an Authentication Token for the Spotify API.
A token is only valid for a few hours. If a 401
error is raised rerun the function to get a new
token.
}
\examples{
getAuthenticationToken(client_id, client_secret_id)
}
