% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_PodcastAPI.R
\name{getEpisodeInformation}
\alias{getEpisodeInformation}
\title{Get Episode Information}
\usage{
getEpisodeInformation(
  episode_id,
  market = "US",
  authentication_token = getAuthenticationToken()
)
}
\arguments{
\item{episode_id}{string: The episode's ID}

\item{market}{string: string: (optional, defaults to US) Returns shows that are available in that market}

\item{authentication_token}{Predefined argument which runs getAuthenticationToken()}
}
\value{
a DataFrame that contains the name and a brief discription of the episode.
}
\description{
Get Episode Information
}
\examples{
getEpisodeInformation('4nRWJ76Tu0ceXJj3uJc4D7')
}
