% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_InfoAPI.R
\name{getRelatedArtists}
\alias{getRelatedArtists}
\title{Get an Artists Related Artists}
\usage{
getRelatedArtists(
  artistId,
  dataframe = TRUE,
  authentication_token = get_authentication_token(CLIENT_ID =
    "7870a259411b4c8b8d2ad173b5a7ed73", CLIENT_SECRET =
    "67ab42b91f224c3682ff8d5b2220f6aa")
)
}
\arguments{
\item{artistId}{The Spotify ID of an artist.}

\item{dataframe}{Boolean. TRUE returns data in a cleaned dataframe, FALSE returns raw JSON object.}

\item{authentication_token}{The users authentication token. Defaults to the value returned by get_authentication_token.}
}
\value{
Ten artists related to the speccified artist.
}
\description{
Get an Artists Related Artists
}
\examples{
getRelatedArtists(artistId = "1Qp56T7n950O3EGMsSl81D", dataframe = TRUE)

}
