% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_InfoAPI.R
\name{getTopSongs}
\alias{getTopSongs}
\title{Get an Artists Top Songs on Spotify}
\usage{
getTopSongs(
  artistId,
  output = "dataframe",
  region = "CA",
  authentication_token = get_authentication_token(CLIENT_ID =
    "7870a259411b4c8b8d2ad173b5a7ed73", CLIENT_SECRET =
    "67ab42b91f224c3682ff8d5b2220f6aa")
)
}
\arguments{
\item{artistId}{The Spotify ID of an artist.}

\item{output}{How to output data. acceptable values are "json", "dataframe", or "graph".}

\item{region}{The two letter code of the market region to check. Default is "CA" for Canada.}

\item{authentication_token}{The users authentication token. Defaults to the value returned by get_authentication_token.}
}
\value{
An object of chosen type containing info on an artists top songs.
}
\description{
Get an Artists Top Songs on Spotify
}
\examples{
getTopSongs("3WPKDlucMsXH6FC1XaclZC", output = "dataframe", region = "CA")
}
