% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrackID.R
\name{getTrackID}
\alias{getTrackID}
\title{Retrieves track IDs users can select from a general string search query}
\usage{
getTrackID(
  track = NA,
  limit = 5,
  authentication_token = getAuthenticationToken()
)
}
\arguments{
\item{track}{The track to search for}

\item{limit}{Optional argument (default = 5) for the number of tracks to return (the greater the number the more likely you track appears in the dataframe)}

\item{authentication_token}{Predefined argument which runs getAuthenticationToken()}
}
\value{
Dataframe containing the track names, artist names, track IDs, and track popularity (0-100)
}
\description{
Retrieves track IDs users can select from a general string search query
}
\examples{
getTrackID('breeze blocks')
}
