% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_TrackRecommender.R
\name{getTrackRecommendations}
\alias{getTrackRecommendations}
\title{Track recommendation function based on seed artists, seed genres, and seed tracks}
\usage{
getTrackRecommendations(
  seed_artists = NA,
  seed_genres = NA,
  seed_tracks = NA,
  authentication_token = get_authentication_token(CLIENT_ID =
    "7870a259411b4c8b8d2ad173b5a7ed73", CLIENT_SECRET =
    "67ab42b91f224c3682ff8d5b2220f6aa"),
  limit = NA,
  market = NA,
  min_acousticness = NA,
  max_acousticness = NA,
  target_acousticness = NA,
  min_danceability = NA,
  max_danceability = NA,
  target_danceability = NA,
  min_duration_ms = NA,
  max_duration_ms = NA,
  target_duration_ms = NA,
  min_energy = NA,
  max_energy = NA,
  target_energy = NA,
  min_instrumentalness = NA,
  max_instrumentalness = NA,
  target_instrumentalness = NA,
  min_key = NA,
  max_key = NA,
  target_key = NA,
  min_liveness = NA,
  max_liveness = NA,
  target_liveness = NA,
  min_loudness = NA,
  max_loudness = NA,
  target_loudness = NA,
  min_mode = NA,
  max_mode = NA,
  target_mode = NA,
  min_popularity = NA,
  max_popularity = NA,
  target_popularity = NA,
  min_speechiness = NA,
  max_speechiness = NA,
  target_speechiness = NA,
  min_tempo = NA,
  max_tempo = NA,
  target_tempo = NA,
  min_time_signature = NA,
  max_time_signature = NA,
  target_time_signature = NA,
  min_valence = NA,
  max_valence = NA,
  target_valence = NA
)
}
\arguments{
\item{seed_artists}{Required artist string or vector of comma separated artist strings}

\item{seed_genres}{Required genre string or vector of comma separated genre strings from spotify's available genre seeds}

\item{seed_tracks}{Required track ID as a string or vector of comma separated track IDs as strings}

\item{authentication_token}{Predefined argument after running get_authentication_token()}

\item{limit}{Optional argument (default = 10) for the number of tracks to recommend}

\item{market}{Optional argument (default = NA) for the ISO 3166-1 alpha-2 country code or the string from auth_token}

\item{min_acousticness}{Optional argument (default = NA) ranges from (0-1)}

\item{max_acousticness}{Optional argument (default = NA) ranges from (0-1)}

\item{target_acousticness}{Optional argument (default = NA) ranges from (0-1)}

\item{min_danceability}{Optional argument (default = NA) ranges from (0-1)}

\item{max_danceability}{Optional argument (default = NA) ranges from (0-1)}

\item{target_danceability}{Optional argument (default = NA) ranges from (0-1)}

\item{min_duration_ms}{Optional argument (default = NA)}

\item{max_duration_ms}{Optional argument (default = NA)}

\item{target_duration_ms}{Optional argument (default = NA)}

\item{min_energy}{Optional argument (default = NA) ranges from (0-1)}

\item{max_energy}{Optional argument (default = NA) ranges from (0-1)}

\item{target_energy}{Optional argument (default = NA) ranges from (0-1)}

\item{min_instrumentalness}{Optional argument (default = NA) ranges from (0-1)}

\item{max_instrumentalness}{Optional argument (default = NA) ranges from (0-1)}

\item{target_instrumentalness}{Optional argument (default = NA) ranges from (0-1)}

\item{min_key}{Optional argument (default = NA)}

\item{max_key}{Optional argument (default = NA)}

\item{target_key}{Optional argument (default = NA)}

\item{min_liveness}{Optional argument (default = NA)}

\item{max_liveness}{Optional argument (default = NA)}

\item{target_liveness}{Optional argument (default = NA)}

\item{min_loudness}{Optional argument (default = NA)}

\item{max_loudness}{Optional argument (default = NA)}

\item{target_loudness}{Optional argument (default = NA)}

\item{min_mode}{Optional argument (default = NA)}

\item{max_mode}{Optional argument (default = NA)}

\item{target_mode}{Optional argument (default = NA)}

\item{min_popularity}{Optional argument (default = NA) ranges from (0-100)}

\item{max_popularity}{Optional argument (default = NA) ranges from (0-100)}

\item{target_popularity}{Optional argument (default = NA) ranges from (0-100)}

\item{min_speechiness}{Optional argument (default = NA) ranges from (0-1)}

\item{max_speechiness}{Optional argument (default = NA) ranges from (0-1)}

\item{target_speechiness}{Optional argument (default = NA) ranges from (0-1)}

\item{min_tempo}{Optional argument (default = NA)}

\item{max_tempo}{Optional argument (default = NA)}

\item{target_tempo}{Optional argument (default = NA)}

\item{min_time_signature}{Optional argument (default = NA)}

\item{max_time_signature}{Optional argument (default = NA)}

\item{target_time_signature}{Optional argument (default = NA)}

\item{min_valence}{Optional argument (default = NA) ranges from (0-1)}

\item{max_valence}{Optional argument (default = NA) ranges from (0-1)}

\item{target_valence}{Optional argument (default = NA) ranges from (0-1)}
}
\value{
A list containing 1) a dataframe with columns for track name, artist name, track id, track popularity, explicit status, and external track link, and 2) a pie chart summarizing the proportion of returned songs which are explicit.
}
\description{
Track recommendation function based on seed artists, seed genres, and seed tracks
}
\examples{
First find song IDs using as an example: get_track_ID('sunday candy')
Then Input Artist, Genre, and track ID parameters to retrieve song recommendations
get_track_recommendations(c('kanye west', 'chance the rapper', 'kendrick'), c('hip hop', 'rap'), '6fTdcGsjxlAD9PSkoPaLMX')
}
