% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrackRecommendations.R
\name{queryAssembler}
\alias{queryAssembler}
\title{Query assembler for the get_track_recommendation() function. This function wasn't designed for use by the user, only back-end processing for structuring the URL query.}
\usage{
queryAssembler(
  seed_artists,
  seed_genres,
  seed_tracks,
  limit,
  market,
  min_acousticness,
  max_acousticness,
  target_acousticness,
  min_danceability,
  max_danceability,
  target_danceability,
  min_duration_ms,
  max_duration_ms,
  target_duration_ms,
  min_energy,
  max_energy,
  target_energy,
  min_instrumentalness,
  max_instrumentalness,
  target_instrumentalness,
  min_key,
  max_key,
  target_key,
  min_liveness,
  max_liveness,
  target_liveness,
  min_loudness,
  max_loudness,
  target_loudness,
  min_mode,
  max_mode,
  target_mode,
  min_popularity,
  max_popularity,
  target_popularity,
  min_speechiness,
  max_speechiness,
  target_speechiness,
  min_tempo,
  max_tempo,
  target_tempo,
  min_time_signature,
  max_time_signature,
  target_time_signature,
  min_valence,
  max_valence,
  target_valence
)
}
\arguments{
\item{seed_artists}{Required artist string or vector of comma separated artist strings}

\item{seed_genres}{Required genre string or vector of comma separated genre strings from spotify's available genre seeds}

\item{seed_tracks}{Required track ID as a string or vector of comma separated track IDs as strings}

\item{limit}{Optional argument (default = 10) for the number of tracks to recommend}

\item{market}{Optional argument (default = NA) for the ISO 3166-1 alpha-2 country code or the string from auth_token}

\item{min_acousticness}{Optional argument (default = NA) ranges from (0-1)}

\item{max_acousticness}{Optional argument (default = NA) ranges from (0-1)}

\item{target_acousticness}{Optional argument (default = NA) ranges from (0-1)}

\item{min_danceability}{Optional argument (default = NA) ranges from (0-1)}

\item{max_danceability}{Optional argument (default = NA) ranges from (0-1)}

\item{target_danceability}{Optional argument (default = NA) ranges from (0-1)}

\item{min_duration_ms}{Optional argument (default = NA)}

\item{max_duration_ms}{Optional argument (default = NA)}

\item{target_duration_ms}{Optional argument (default = NA)}

\item{min_energy}{Optional argument (default = NA) ranges from (0-1)}

\item{max_energy}{Optional argument (default = NA) ranges from (0-1)}

\item{target_energy}{Optional argument (default = NA) ranges from (0-1)}

\item{min_instrumentalness}{Optional argument (default = NA) ranges from (0-1)}

\item{max_instrumentalness}{Optional argument (default = NA) ranges from (0-1)}

\item{target_instrumentalness}{Optional argument (default = NA) ranges from (0-1)}

\item{min_key}{Optional argument (default = NA)}

\item{max_key}{Optional argument (default = NA)}

\item{target_key}{Optional argument (default = NA)}

\item{min_liveness}{Optional argument (default = NA)}

\item{max_liveness}{Optional argument (default = NA)}

\item{target_liveness}{Optional argument (default = NA)}

\item{min_loudness}{Optional argument (default = NA)}

\item{max_loudness}{Optional argument (default = NA)}

\item{target_loudness}{Optional argument (default = NA)}

\item{min_mode}{Optional argument (default = NA)}

\item{max_mode}{Optional argument (default = NA)}

\item{target_mode}{Optional argument (default = NA)}

\item{min_popularity}{Optional argument (default = NA) ranges from (0-100)}

\item{max_popularity}{Optional argument (default = NA) ranges from (0-100)}

\item{target_popularity}{Optional argument (default = NA) ranges from (0-100)}

\item{min_speechiness}{Optional argument (default = NA) ranges from (0-1)}

\item{max_speechiness}{Optional argument (default = NA) ranges from (0-1)}

\item{target_speechiness}{Optional argument (default = NA) ranges from (0-1)}

\item{min_tempo}{Optional argument (default = NA)}

\item{max_tempo}{Optional argument (default = NA)}

\item{target_tempo}{Optional argument (default = NA)}

\item{min_time_signature}{Optional argument (default = NA)}

\item{max_time_signature}{Optional argument (default = NA)}

\item{target_time_signature}{Optional argument (default = NA)}

\item{min_valence}{Optional argument (default = NA) ranges from (0-1)}

\item{max_valence}{Optional argument (default = NA) ranges from (0-1)}

\item{target_valence}{Optional argument (default = NA) ranges from (0-1)}
}
\value{
A formatted string query to use in track recommendation API calls
}
\description{
Query assembler for the get_track_recommendation() function. This function wasn't designed for use by the user, only back-end processing for structuring the URL query.
}
