% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyPodcastAPI.R
\name{searchForPodcast}
\alias{searchForPodcast}
\title{Title}
\usage{
searchForPodcast(
  keywords,
  language = "en",
  market = "US",
  explicit = TRUE,
  limit = 5
)
}
\arguments{
\item{keywords}{}

\item{language}{}

\item{market}{}

\item{explicit}{}

\item{limit}{}
}
\value{

}
\description{
Title
}
\examples{

}
