% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EricsFunctions.R
\name{get_authentication_token}
\alias{get_authentication_token}
\title{Get an Authentication Token for the Spotify API}
\usage{
get_authentication_token(client_id, client_secret_id)
}
\arguments{
\item{client_id}{Your client id. Can be found at https://developer.spotify.com/dashboard}

\item{client_secret_id}{Similar to client id}
}
\value{
an authentication token, saved to global environments as auth_token
}
\description{
Queries the API for a token and sets the token as a global variable
}
\details{
A token will only remain valid for a few hours. If the package gives 401 errors, rerun this function.
}
\examples{
get_authentication_token(client_id, client_secret_id)

}
