% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EricsFunctions.R
\name{getAudioFeatures}
\alias{getAudioFeatures}
\title{Get Audio Features of a Song}
\usage{
getAudioFeatures(
  songId,
  output = "dataframe",
  authentication_token = auth_token
)
}
\arguments{
\item{songId}{The Id of a song to search}

\item{output}{Output type. "json", "dataframe", or "graph"}

\item{authentication_token}{the users authentication token for the Spotify API. Defaults to the value saved when get_authentication_token is run.}
}
\value{
A json object, dataframe, or graph describing song features
}
\description{
Spotify automatically generates a number of audio features for each song.
}
\examples{
getAudioFeatures("68ngtC3pGiTjXcFwxYCJ7Z", output = "graph")
}
