% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAuthenticationToken.R
\name{getAuthenticationToken}
\alias{getAuthenticationToken}
\title{Get an API authentication token.}
\usage{
getAuthenticationToken(
  CLIENT_ID = "7870a259411b4c8b8d2ad173b5a7ed73",
  CLIENT_SECRET = "67ab42b91f224c3682ff8d5b2220f6aa"
)
}
\arguments{
\item{CLIENT_ID}{A Spotify client ID which the user needs to acquire prior to using the package.}

\item{CLIENT_SECRET}{A Spotify client secret key which the user needs to acquire prior to using the package.}
}
\value{
auth_token
}
\description{
Retrieves an Authentication Token for the Spotify API.
The function is called when the API is called so
expired tokens are refreshed when needed automatically
}
\examples{
getAuthenticationToken(CLIENT_ID, CLIENT_SECRET)
}
