% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_PodcastAPI.R
\name{getPodcastID}
\alias{getPodcastID}
\title{Get Show ID}
\usage{
getPodcastID(
  query,
  market = "US",
  authentication_token = getAuthenticationToken()
)
}
\arguments{
\item{query}{string: The show to search for}

\item{market}{string: (optional, defaults to US) A string returning shows that are available in that market}
}
\value{
A string containing the show's ID
}
\description{
Retrieves a podcast ID from a general name search query
}
\examples{
getPodcastID('Philosophize This!', market='US')
getPodcastID('Philosophize This!')
}
