% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EricsFunctions.R
\name{getRelatedArtists}
\alias{getRelatedArtists}
\title{Get an Artists Related Artists}
\usage{
getRelatedArtists(
  artistId,
  dataframe = TRUE,
  authentication_token = auth_token
)
}
\arguments{
\item{artistId}{The Spotify ID of an artist.}

\item{dataframe}{Boolean. TRUE returns data in a cleaned dataframe, FALSE returns raw JSON object.}

\item{authentication_token}{The users authentication token. Defaults to the value returned by get_authentication_token.}
}
\value{
Ten artists related to the speccified artist.
}
\description{
Get an Artists Related Artists
}
\examples{
getRelatedArtists(artistId = "1Qp56T7n950O3EGMsSl81D", dataframe = TRUE)

}
