% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_InfoAPI.R
\name{getTopSongs}
\alias{getTopSongs}
\title{Get an Artists Top Songs on Spotify}
\usage{
getTopSongs(
  artistId,
  output = "dataframe",
  region = "CA",
  authentication_token = getAuthenticationToken()
)
}
\arguments{
\item{artistId}{The Spotify ID of an artist.}

\item{output}{How to output data. acceptable values are "json", "dataframe", or "graph".}

\item{region}{The two letter code of the market region to check. Default is "CA" for Canada.}

\item{authentication_token}{Predefined argument which runs getAuthenticationToken()}
}
\value{
An object of chosen type containing info on an artists top songs.
}
\description{
Get an Artists Top Songs on Spotify
}
\examples{
getTopSongs("3WPKDlucMsXH6FC1XaclZC", output = "dataframe", region = "CA")
}
