% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_PodcastAPI.R
\name{searchForPodcast}
\alias{searchForPodcast}
\title{Search for a new podcast}
\usage{
searchForPodcast(
  keywords,
  language = "en",
  market = "US",
  explicit = TRUE,
  limit = 5,
  authentication_token = getAuthenticationToken()
)
}
\arguments{
\item{keywords}{string: The search query}

\item{language}{string: (optional, defaults to ENGLISH)}

\item{market}{string: (optional, defaults to US) Returns shows that are available in that market}

\item{explicit}{logical: (optional, defaults to TRUE) To enable the filter set explict to FALSE}

\item{limit}{integer: (optional, defaults to 5, min = 1, max = 50) Number of shows to be returned}
}
\value{
Dataframe containing the podcast name, publisher, language, explicit content filter and podcast ID
}
\description{
Search for a new podcast
}
\examples{
searchForPodcast('History')
searchForPodcast('History', language='es')
searchForPodcast('History', language='es', market='ES')
searchForPodcast('History', language='es', market='ES', explicit=FALSE)
searchForPodcast('History', language='es', market='ES', explicit=FALSE, limit=10)
}
